Add versioning scheme and importing monosat if already there

+++ build.sbt
@@ -81,6 +81,8 @@ lazy val dockerSettings = Seq(
     val minimalBuildSBT = writeMinimalBuildSBT.value
     val artifactTargetPath = s"/home/user/code/lib/${artifact.name}"
     val base = (Compile / baseDirectory).value
+    val monosatLib: File = base / "lib" / "monosat.jar"
+    val monosatDynlib: File = base / "binlib" / "monosat.so"
     new Dockerfile {
       from("openjdk:8")
       customInstruction("RUN", "apt-get update && apt-get --fix-missing update && apt-get install -y graphviz gnupg libgmp3-dev make cmake build-essential zlib1g-dev")
@@ -100,12 +102,17 @@ lazy val dockerSettings = Seq(
       customInstruction("RUN", "mkdir -p /home/user/code/src/main/scala/onera/pmlanalyzer/views/interference")
       customInstruction("RUN", "mkdir -p /home/user/code/src/test")
       workDir("/home/user")
-      customInstruction("RUN", "git clone https://github.com/sambayless/monosat.git")
-      workDir("/home/user/monosat")
-      customInstruction("RUN", "cmake -DJAVA=ON .")
-      customInstruction("RUN", "make")
-      customInstruction("RUN", "cp libmonosat.so /home/user/code/binlib")
-      workDir("/home/user/code")
+      if(monosatLib.exists() && monosatDynlib.exists()){
+        copy(monosatLib,"code/lib")
+        copy(monosatDynlib,"code/binlib")
+      } else {
+        customInstruction("RUN", "git clone https://github.com/sambayless/monosat.git")
+        workDir("/home/user/monosat")
+        customInstruction("RUN", "cmake -DJAVA=ON .")
+        customInstruction("RUN", "make")
+        customInstruction("RUN", "cp libmonosat.so /home/user/code/binlib")
+      }
+        workDir("/home/user/code")
       for ((to, from) <- modelCode.value)
         copy(from, to)
       copy((Compile / doc / target).value, "doc")
@@ -174,6 +181,7 @@ lazy val commonSettings = Seq(
   ),
   pomIncludeRepository := { _ => false },
   crossPaths := false,
+  versionScheme := Some("early-semver"),
   scalaVersion := "3.3.5",
   sbtVersion := "1.11.2",
   scalafixOnCompile := true,
